Fixed a few more implicit casts, removed a few unused variables

+++ Grbl_Esp32/src/Config.h
@@ -163,7 +163,7 @@ const int TOOL_LENGTH_OFFSET_AXIS = Z_AXIS;  // Default z-axis. Valid values are
 // limits or angle between neighboring block line move directions. This is useful for machines that can't
 // tolerate the tool dwelling for a split second, i.e. 3d printers or laser cutters. If used, this value
 // should not be much greater than zero or to the minimum value necessary for the machine to work.
-const double MINIMUM_JUNCTION_SPEED = 0.0;  // (mm/min)
+const float MINIMUM_JUNCTION_SPEED = 0.0f;  // (mm/min)
 
 // Sets the minimum feed rate the planner will allow. Any value below it will be set to this minimum
 // value. This also ensures that a planned motion always completes and accounts for any floating-point
